+++ test/fixtures/app/installers/thumbnail_installer.php
@@ -6,7 +6,7 @@ class ThumbnailInstaller extends AkInstaller
     {
         $this->createTable('thumbnails', '
         id integer(11) auto increment primary key,
-        photo_id integer(10),
+        photo_id integer,
         owner string (40) default \'Picture\',
         caption string (200)
         '
